Refactoring & style change pin view
