+++ app/app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".fieldbook.FieldBook"></activity>
+        <activity android:name=".FieldBook"></activity>
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -25,7 +25,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".ui.TouchParent" />
+        <activity android:name=".gestureDetection.TouchParent" />
         <activity
             android:name=".GeoMap"
             android:label="@string/title_activity_geo_map"
