+++ app/app/src/main/java/com/uu_uce/gestureDetection/DoubleTapper.kt
@@ -1,15 +1,13 @@
-package com.uu_uce.ui
+package com.uu_uce.gestureDetection
 
-import android.app.Activity
+import android.content.Context
 import android.view.GestureDetector
 import android.view.MotionEvent
-import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.GestureDetectorCompat
-import com.uu_uce.shapefiles.p2
 
-class SingleTapper(
-    val parent: AppCompatActivity,
-    val action: (p2, Activity) -> Unit)
+class DoubleTapper(
+    parent: Context,
+    var action: () -> Unit)
     : TouchChild,
     GestureDetector.OnGestureListener,
     GestureDetector.OnDoubleTapListener{
@@ -24,13 +22,13 @@ class SingleTapper(
         gestureDetector?.onTouchEvent(event)
     }
 
-    override fun onSingleTapConfirmed(e: MotionEvent?): Boolean {
-        action(Pair((e?.x ?: 0.0f).toDouble(), (e?.y ?: 0.0f).toDouble()), parent)
+    override fun onDoubleTap(e: MotionEvent?): Boolean {
+        action()
         return true
     }
 
-    override fun onDoubleTap(e: MotionEvent?): Boolean { return true }
     override fun onDoubleTapEvent(e: MotionEvent?): Boolean { return true }
+    override fun onSingleTapConfirmed(e: MotionEvent?): Boolean { return true }
     override fun onShowPress(e: MotionEvent?) {}
     override fun onSingleTapUp(e: MotionEvent?): Boolean { return true }
     override fun onDown(e: MotionEvent?): Boolean { return false }
