+++ app/app/src/main/res/layout/pin_content_view.xml
@@ -4,7 +4,8 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="#33000000"
+    android:background="@color/colorUU"
+    android:padding="5dp"
     android:gravity="center">
 
     <androidx.cardview.widget.CardView
